refactor(external_sync): split 2932-line file into mod/types/normalize/merge/cursor/removals + tests
services/external_sync was the single biggest unsplit file in the repo
after #62 PR E shipped. Production code now lives in 6 cohesive
siblings under services/external_sync/:

  mod.rs       712  Tick orchestration: tick_once / tick_once_or_busy /
                    tick_once_inner, sync_anilist + sync_mal AL/MAL
                    fetch wrappers, EXTERNAL_SYNC_LOCK, the auth-
                    rejection taxonomy, post-merge bulk-pass spawn,
                    has_linked_account / minutes_since_last_run.
  types.rs     232  NormalizedStatus + impl, SyncEntry, MergeOutcome
                    (incl. merge_pass), NewArtworkSpec, RemovalReport,
                    the internal MergeAction enum the merge bodies
                    return per-row.
  normalize.rs 158  entries_from_anilist / entries_from_mal,
                    monitor_mode_for + import_status, anibridge MAL→AL
                    resolve_mal_anilist_ids.
  merge.rs     440  merge_into_library, the per-entry AL + Jikan merge
                    bodies, the stamp_* cross-cutting helpers,
                    merge_jikan_fallback_entries.
  removals.rs   61  detect_removals (full-resync-only downgrade pass).
  cursor.rs     55  should_full_resync + drop_entries_before_cursor +
                    FULL_RESYNC_INTERVAL_SECS.

Tests stay in a single tests/mod.rs (1366 LoC, just under the ~1500
LoC topic-split threshold from CLAUDE.md). The 43 tests there cover
the same behaviors as before — no test bodies changed, just relocated
+ given the explicit imports (series, MonitorMode, anibridge, anilist,
jikan, mal, series_custom_lists, HashMap) that the prior `use super::*;`
implicitly inherited from the bigger mod.rs.

Public API is identical — `pub use` re-exports in mod.rs cover
everything main.rs / handlers::oauth / tests/external_sync_e2e.rs
already reach for. is_auth_rejection + EXTERNAL_SYNC_LOCK +
current_unix_ts are now `pub(crate)` because the moved tests
reference them.

Verify: 1233 unit + 7 PR0 + 1 e2e tests all green, fmt + clippy
-D warnings clean.
